+++ tests/3xx/309_is_empty.sh
@@ -5,7 +5,7 @@
 assert_func is_empty "${TEST_FAIL}" "${TEST_EMPTY}"
 assert_func is_empty "${TEST_FAIL}" "${TEST_EMPTY}" '/tmp' 'bla'
 
-assert_func is_empty "${TEST_OK}" "${TEST_EMPTY}" ''
+assert_func is_empty "${TEST_FAIL}" "${TEST_EMPTY}" ''
 assert_func is_empty "${TEST_FAIL}" "${TEST_EMPTY}" 'foo'
 
 declare -g -a TESTARY=()